+++ clarity-types/src/lib.rs
@@ -37,11 +37,11 @@ use stacks_common::types::StacksEpochId;
 pub use types::Value;
 
 /// Max call stack depth for Epoch 3.4+.
-const MAX_CALL_STACK_DEPTH: usize = 128;
+const MAX_CALL_STACK_DEPTH: u64 = 128;
 /// Max call stack depth for pre‑3.4 epochs.
-const MAX_CALL_STACK_DEPTH_LEGACY: usize = 64;
+const MAX_CALL_STACK_DEPTH_LEGACY: u64 = 64;
 
-pub fn max_call_stack_depth_for_epoch(epoch_id: StacksEpochId) -> usize {
+pub fn max_call_stack_depth_for_epoch(epoch_id: StacksEpochId) -> u64 {
     if epoch_id >= StacksEpochId::Epoch34 {
         MAX_CALL_STACK_DEPTH
     } else {
